setup PUT route poccibly needs changes

+++ pages/api/machines/[id].js
@@ -5,6 +5,19 @@ export default async function handler(request, response) {
   await dbConnect();
   const machineId = request.query.id;
 
+  if (request.method === "PUT") {
+    const machineUpdateData = request.body;
+    const machine = await Machine.findByIdAndUpdate(
+      machineId,
+      machineUpdateData
+    );
+
+    if (!machine) {
+      return response.status(404).json({ status: "Machine not found" });
+    }
+    return response.status(200).json({ status: "Machine updated" });
+  }
+
   if (request.method === "DELETE") {
     const machine = await Machine.findByIdAndDelete(machineId);
